requirements.txt
FastAPI
mysql-connector-python
